Support sync messages in multi-shard setup
Process sync message correctly in cases when there is no
active backend connections to any shard, instead of replying with
error 'too complex to parse'. This fixed the case when sync message is
issued right after parse message, or similar, in which router fill fail to
'route' sync to shard, because this is un-routabale messafe.